cloud_functions/data_loader/requirements.txt
cachetools==5.2.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4"
cloudevents==1.9.0 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4"
flask==2.2.2 ; python_version >= "3.10" and python_version < "4"
functions-framework==3.3.0 ; python_version >= "3.10" and python_version < "4"
google-api-core==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.16.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery==3.4.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version < "4.0" and python_version >= "3.10"
grpcio-status==1.51.1 ; python_version < "4.0" and python_version >= "3.10"
grpcio==1.51.1 ; python_version < "4.0" and python_version >= "3.10"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4"
numpy==1.24.1 ; python_version < "4.0" and python_version >= "3.10"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.22.2 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.12 ; python_version < "4.0" and python_version >= "3.10"
pyarrow==10.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==66.1.1 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
watchdog==2.2.1 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.2.2 ; python_version >= "3.10" and python_version < "4"
